pyproject.toml
[project]
name = "honey-commander"
version = "0.1.2"
description = "꿀 커맨더 - Cross-platform file explorer with 꿀뷰-style image viewer"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=10.0.0",
    "psd-tools>=1.12.1",
    "pyside6>=6.8.0",
    "rarfile>=4.2",
    "send2trash>=1.8.3",
    "pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'",
    "py7zr>=1.1.0",
    "pygments>=2.19.2",
]

[project.scripts]
commander = "commander.__main__:main"

[project.gui-scripts]
commander-gui = "commander.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "pytest-cov>=7.0.0",
]
build = [
    "pyinstaller>=6.0.0",
]
extra-formats = [
    "pillow-heif>=0.16.0",
    "rawpy>=0.19.0",
    "OpenEXR>=3.2.0",
    "numpy>=1.24.0",
]
network = [
    "keyring>=24.0.0",
    "paramiko>=3.0.0",
    "smbprotocol>=1.10.0",
]
viewer-3d = [
    "pyvista>=0.43.0",
    "pyvistaqt>=0.11.0",
    "trimesh>=4.0.0",
    "pyassimp>=4.1.0",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/commander"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
ignore = ["E402"]  # Allow imports after setup code

[tool.ruff.lint.per-file-ignores]
"src/commander/core/image_loader.py" = ["F401"]  # pillow_avif import for side effects
"src/commander/views/file_list/drop_views.py" = ["E402"]  # Qt import after mixin class

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
